generate random names for workspaces
