+++ components/profile/interfaces.ts
@@ -4,56 +4,3 @@ export type Social = {
     discordUsername?: string;
     linkedinURL?: string;
 };
-
-export const CRYPTO_WORD_LIST = [
-  'airdrop',
-  'alpha',
-  'apein',
-  'bearish',
-  'blockchain',
-  'bluechip',
-  'bullish',
-  'coinbase',
-  'cryptowallet',
-  'cryptowalletaddress',
-  'decentralized',
-  'dao',
-  'dapp',
-  'defi',
-  'diamondhands',
-  'doxxed',
-  'drop',
-  'dyor',
-  'doteth',
-  'ethereum',
-  'flip',
-  'floor',
-  'floorprice',
-  'floorsweep',
-  'fomo',
-  'fud',
-  'gasfee',
-  'gm',
-  'hodl',
-  'lfg',
-  'marketplace',
-  'metamask',
-  'metaverse',
-  'minting',
-  'mooning',
-  'ngmi',
-  'opensea',
-  'paperhands',
-  'pfp',
-  'polygon',
-  'pumpanddump',
-  'p2e',
-  'roadmap',
-  'rugpull',
-  'sharding',
-  'solana',
-  'smartcontract',
-  'staking',
-  'tothemoon',
-  'utility'
-];
